// src/lib/generated/types.gen.ts
export type UserResponse = {
    id: number;
    email: string;
    nickname: string;
    profileImageUrl?: string;
    role: 'USER' | 'ADMIN' | 'GUIDE' | 'PENDING';
};